RpcValue: Cast bools to ints as 0
Casting bools to ints never makes sense. The reason I found is in the
checkAlarms algorithm, where if you put a boolean as `value`, it could later be
interpreted as a bitfield, because of this implicit conversion. It'll still do
it now, but at least it won't result in `true` which will stop it from deducing
invalid alarms.